Publish a new version of the app. Installation procedure will uninstall old version of the app (with backing up the local DB)
Figured out that when the new version of the app was published the old
version of the app was not manually uninstall and as a result there were
two versions of the application on the tablets. The users were opening
the old version and a confusion was made.

- In this installation process any old version of the app is uninstalled
and only one exists on the tablet. The local DB is backed-up from the
old version of the application and restored.
- The settings that are saved on the application (links, logo, name of
Uni) are preserved when new version of the application will be
installed.
- Styling bug fixes. Old custom styling has been removed and Bootstrap
is used all over.
- Added more error handling.
- 'Actions' tab on the list of forms has a new functionality when the
app is used on a mobile based device (with small screen size).

- Removed old installation files