Feb 2017 updates
- The feedback box fills the column’s width.
- Moodle url path: The url input texts have been split up into 2
sections. Host name and the according webpage/api call. There is only
one input text to determine the moodle host name url. No need to type it
3 times. There are 3 inputs to type in the webpage/api call. All these
fields will be populated by default but the user must click the ‘Verify’
button and save.
- Barcode: The ‘barcode’ checkbox is not mandatory any more.
- Fixed the user interface on the yellow preview tab in exam view.
- Display App version on navigation bar.
- Fixed ‘forms’ ordering and remember the filter settings when going
back and forth between the forms table page and opening a form for
assessment.
- Fixed ‘downloaded forms’ ordering and remember the filter settings.
The user can order by ‘Course Name’ or ‘ID’ and then by ‘title’.
- Faster loading of the ‘downloaded forms’ and ‘forms’ tables.
- Faster deletion of form folder.
- Faster downloading/populating exam.
- Added 2 new columns to the downloaded forms table. The user can now
see the course name of each exam.
If there are already forms on the tablet (after the latest update) the
user must login and 'SYNC' the table.
- The api has changed in order to synchronise the exams’ course name and
id. Also I’ve made some more alterations in order to download the forms
faster.
When the time comes to test the new version of the app you must remember
to paste the new ‘json_produce.php’ file to the server. The name of the
updated file is ‘json_produce_v2.php’. Please make sure to set the
correct path in the configuration before start testing.